pyproject.toml
[project]
name = "student-schedule"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=2.3.0, <2.4",
    "django>=5.1.3, <5.2",
    "psycopg[binary]>=3.2.3, <3.3",
    "pydantic-settings>=2.6.1, <2.7",
    "pydantic>=2.10.1, <2.11",
    "djangorestframework>=3.15.2, <3.16",
    "django-filter>=24.3",
]

[dependency-groups]
dev = [
    "ipython>=8.29.0",
    "ruff>=0.8.0",
]
